refactor(rec form): reorganize file
