+++ apps/recnet/src/app/MobileNavigator.tsx
@@ -13,8 +13,8 @@ import { useState } from "react";
 import { toast } from "sonner";
 
 import { trpc } from "@recnet/recnet-web/app/_trpc/client";
-import { RecArticleForm } from "@recnet/recnet-web/components/RecArticleForm";
 import { SkeletonText, Skeleton } from "@recnet/recnet-web/components/Skeleton";
+import { RecArticleForm } from "@recnet/recnet-web/components/rec/RecArticleForm";
 import { UserRole } from "@recnet/recnet-web/constant";
 import { useGoogleLogin } from "@recnet/recnet-web/firebase/auth";
 import { cn } from "@recnet/recnet-web/utils/cn";
